Add alert message to front end and edit according to api
