+++ Description for Status Codes.md
@@ -23,3 +23,6 @@ The requested resource is not found.
 ## 5000 Server Error
 **5000 Internal Server Error**
 The server encountered an unexpected condition in fulfilling the request.
+
+**5001 Duplicated Entry**
+The database encountered an unexpected condition in fulfilling the request.
