+++ restaurant-remote-reception-app/src/app/app.component.html
@@ -24,5 +24,5 @@
     </div>
   </div>
 </nav>
-
+<app-alert></app-alert>
 <router-outlet></router-outlet>
