+++ restaurant-remote-reception-app/src/app/components/alert/alert.component.html
@@ -1,0 +1,1 @@
+<div *ngIf="message" [ngClass]="{ 'alert': message, 'alert-success': message.type === 'success', 'alert-danger': message.type === 'error' }">{{message.text}}</div>
