+++ restaurant-remote-reception-app/src/app/components/alert/alert.component.ts
@@ -1,0 +1,18 @@
+import { Component, OnInit } from '@angular/core';
+import {AlertService} from '../../services/alert.service';
+
+@Component({
+  selector: 'app-alert',
+  templateUrl: './alert.component.html',
+  styleUrls: ['./alert.component.css']
+})
+export class AlertComponent implements OnInit {
+  message: any;
+
+  constructor(private alertService: AlertService) { }
+
+  ngOnInit() {
+    this.alertService.getMessage().subscribe(message => { this.message = message; });
+  }
+
+}
