+++ restaurant-remote-reception-app/src/app/components/restaurant-list/restaurant-list.component.html
