+++ restaurant-remote-reception-app/src/app/components/shopping-center-list/shopping-center-list.component.css
