+++ restaurant-remote-reception-app/src/app/components/signup/signup.component.html
@@ -18,7 +18,7 @@
         <div *ngIf="signupForm.submitted && inputPassword.invalid" class="invalid-feedback">パスワードの入力は必要です。</div>
         <small id="passwordHelp" class="form-text text-muted">6文字以上で入力してください。</small>
       </div>
-      <button type="submit" class="btn btn-primary">登録</button> <img *ngIf="isLoading" src="../../assets/icon_loader.gif">
+      <button type="submit" class="btn btn-primary">登録</button> <img *ngIf="isLoading" src="../../../assets/icon_loader.gif">
     </form>
   </div>
 </div>
