+++ restaurant-remote-reception-app/src/app/components/ticket-display-panel/ticket-display-panel.component.css
