+++ restaurant-remote-reception-app/src/app/components/ticket-display-panel/ticket-display-panel.component.ts
@@ -1,8 +1,9 @@
 import { Component, OnInit, OnDestroy, Input } from '@angular/core';
-import { TicketService} from '../services/ticket.service';
-import { AppUtils} from '../app-common';
+import { TicketService} from '../../services/ticket.service';
+import { AppUtils} from '../../app-common';
 import {interval} from 'rxjs';
-import {ITicketCurrentCount, TICKET_TYPES} from '../models/ticket.model';
+import {ITicketCurrentCount, TICKET_TYPES} from '../../models/ticket.model';
+import {AlertService} from '../../services/alert.service';
 
 const intervalCounter = interval(10000);
 
@@ -16,7 +17,8 @@ export class TicketDisplayPanelComponent implements OnInit, OnDestroy {
   ticketCurrentCounts: ITicketCurrentCount[];
   alive: boolean;
 
-  constructor(private ticketService: TicketService) {
+  constructor(private ticketService: TicketService,
+              private alertService: AlertService) {
     this.alive = true;
   }
 
@@ -43,7 +45,17 @@ export class TicketDisplayPanelComponent implements OnInit, OnDestroy {
         this.ticketCurrentCounts = res;
       },
       err => {
-        AppUtils.handleError(err);
+        if (err.error instanceof Error) { // browser error
+          this.alertService.error(0, err.error.message);
+        } else if (err.error.message == null) { // non-customised error
+          this.alertService.error(err.status, err.statusText);
+        } else if (err.error.status_code >= 5000 ) { // server error with message not to be shown on UI
+          this.alertService.error(err.error.status_code, err.statusText);
+        } else if (err.error.status_code === 4011) {
+          this.alertService.error(err.error.status_code, 'Please login');
+        } else {
+          this.alertService.error(err.error.status_code, err.error.message);
+        }
       }
     );
   }
