+++ restaurant-remote-reception-app/src/app/components/ticket-reservation/ticket-reservation.component.html
@@ -22,7 +22,7 @@
         </div>
 
       </div>
-      <button type="submit" class="btn btn-primary">確認</button> <img *ngIf="isLoading" src="../../assets/icon_loader.gif">
+      <button type="submit" class="btn btn-primary">確認</button> <img *ngIf="isLoading" src="../../../assets/icon_loader.gif">
     </form>
   </div>
 </div>
