+++ restaurant-remote-reception-app/src/app/components/ticket-reservation/ticket-reservation.component.ts
@@ -1,0 +1,110 @@
+import { Component, OnInit } from '@angular/core';
+import {ActivatedRoute, Router} from '@angular/router';
+import { IRestaurant, RestaurantService } from '../../services/restaurant.service';
+import {IUser} from '../../models/user.model';
+import {Ticket} from '../../models/ticket.model';
+import {TicketService} from '../../services/ticket.service';
+import {UserService} from '../../services/user.service';
+import {AlertService} from '../../services/alert.service';
+
+@Component({
+  selector: 'app-ticker-reservation',
+  templateUrl: './ticket-reservation.component.html',
+  styleUrls: ['./ticket-reservation.component.css']
+})
+export class TicketReservationComponent implements OnInit {
+  restaurant: IRestaurant;
+  user: IUser;
+  ticket = new Ticket(0, 0, 1, 'Active');
+  isLoading = false;
+
+  constructor(
+    private route: ActivatedRoute,
+    private router: Router,
+    private restaurantService: RestaurantService,
+    private userService: UserService,
+    private ticketService: TicketService,
+    private alertService: AlertService
+  ) { }
+
+  ngOnInit() {
+    this.ticket.restaurant_id = parseInt(this.route.snapshot.paramMap.get('restaurantId'), 10);
+    this.getRestaurantInfo();
+    this.getUserInfo();
+  }
+
+  onSubmit() {
+    this.isLoading = true;
+    this.reserveTicket();
+  }
+
+  private reserveTicket() {
+    const ticketForm = JSON.stringify(this.ticket);
+    this.ticketService.create(ticketForm).subscribe(
+      (res: Response) => {
+        this.router.navigate(['/shopping-centers', this.restaurant.shopping_center_id]);
+      },
+      err => {
+        if (err.error instanceof Error) { // browser error
+          this.alertService.error(0, err.error.message);
+        } else if (err.error.message == null) { // non-customised error
+          this.alertService.error(err.status, err.statusText);
+        } else if (err.error.status_code >= 5000 ) { // server error with message not to be shown on UI
+          this.alertService.error(err.error.status_code, err.statusText);
+        } else if (err.error.status_code === 4011) {
+          this.alertService.error(err.error.status_code, 'Please login');
+        } else if (err.error.status_code === 4000) {
+          this.alertService.error(err.error.status_code, 'Invalid input');
+        } else {
+          this.alertService.error(err.error.status_code, err.error.message);
+        }
+        this.isLoading = false;
+      }
+    );
+  }
+
+  private getUserInfo() {
+      this.userService.getCurrentUserInfo().subscribe(
+        (res: IUser) => {
+          this.user = res;
+          this.ticket.created_by_id = this.user.id;
+        },
+        err => {
+          if (err.error instanceof Error) { // browser error
+            this.alertService.error(0, err.error.message);
+          } else if (err.error.message == null) { // non-customised error
+            this.alertService.error(err.status, err.statusText);
+          } else if (err.error.status_code >= 5000 ) { // server error with message not to be shown on UI
+            this.alertService.error(err.error.status_code, err.statusText);
+          } else if (err.error.status_code === 4011) {
+            this.alertService.error(err.error.status_code, 'Please login');
+          } else {
+            this.alertService.error(err.error.status_code, err.error.message);
+          }
+        }
+      );
+    }
+
+  private getRestaurantInfo() {
+    this.restaurantService.getInfo(this.ticket.restaurant_id.toString()).subscribe(
+      (res: IRestaurant) => {
+        this.restaurant = res;
+      },
+      err => {
+        if (err.error instanceof Error) { // browser error
+          this.alertService.error(0, err.error.message);
+        } else if (err.error.message == null) { // non-customised error
+          this.alertService.error(err.status, err.statusText);
+        } else if (err.error.status_code >= 5000 ) { // server error with message not to be shown on UI
+          this.alertService.error(err.error.status_code, err.statusText);
+        } else if (err.error.status_code === 4011) {
+          this.alertService.error(err.error.status_code, 'Please login');
+        } else if (err.error.status_code === 4040) {
+          this.alertService.error(err.error.status_code, 'Restaurant not found');
+        } else {
+          this.alertService.error(err.error.status_code, err.error.message);
+        }
+      }
+    );
+  }
+}
