+++ restaurant-remote-reception-app/src/app/components/user-login/user-login.component.html
@@ -13,7 +13,7 @@
         <small id="passwordHelp" class="form-text text-muted">6文字以上で入力してください。</small>
         <div *ngIf="loginForm.submitted && inputPassword.invalid" class="invalid-feedback">パスワードの入力は必要です。</div>
       </div>
-      <button type="submit" class="btn btn-primary">ログイン</button> <img *ngIf="isLoading" src="../../assets/icon_loader.gif">
+      <button type="submit" class="btn btn-primary">ログイン</button> <img *ngIf="isLoading" src="../../../assets/icon_loader.gif">
       <button type="button" class="btn btn-link" routerLink="/signup">登録</button>
     </form>
   </div>
