+++ restaurant-remote-reception-app/src/app/services/alert.service.ts
@@ -1,0 +1,44 @@
+import { Injectable } from '@angular/core';
+import {Observable, Subject} from 'rxjs';
+import {NavigationStart, Router} from '@angular/router';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class AlertService {
+  private subject = new Subject<any>();
+  private keepAfterNavigationChange = false;
+
+  constructor(private router: Router) {
+    router.events.subscribe(event => {
+      if (event instanceof NavigationStart) {
+        if (this.keepAfterNavigationChange) {
+          // only keep for a single location change
+          this.keepAfterNavigationChange = false;
+        } else {
+          // clear alert
+          this.subject.next();
+        }
+      }
+    });
+  }
+
+  success(message: string, keepAfterNavigationChange = false) {
+    this.keepAfterNavigationChange = keepAfterNavigationChange;
+    this.subject.next({ type: 'success', text: message });
+  }
+
+  error(code = 0, message: string, keepAfterNavigationChange = false) {
+    this.keepAfterNavigationChange = keepAfterNavigationChange;
+    if (code > 0) {
+      this.subject.next({type: 'error', text: 'Error (' + code + ') ' + message});
+    } else {
+      this.subject.next({type: 'error', text: 'Error ' + message});
+    }
+  }
+
+  getMessage(): Observable<any> {
+    return this.subject.asObservable();
+  }
+
+}
