+++ restaurant-remote-reception/app/controllers/ShoppingCenterController.scala
@@ -23,7 +23,12 @@ class ShoppingCenterController @Inject()(val dbConfigProvider: DatabaseConfigPro
         db.run(Users.filter(t => t.userId === sessionUserId.toInt).result.headOption).flatMap {
             case Some(_) =>
                 db.run(ShoppingCenters.sortBy(t => t.shoppingCenterId).result)
-                        .map(shoppingCenters => Ok(Json.toJson(shoppingCenters)))
+                    .map{
+                        case shoppingCenters if shoppingCenters.nonEmpty =>
+                            Ok(Json.toJson(shoppingCenters))
+                        case shoppingCenters if shoppingCenters.isEmpty =>
+                            NotFound(Json.toJson(StatusResponse(StatusCode.RESOURCE_NOT_FOUND.code, StatusCode.RESOURCE_NOT_FOUND.message)))
+                    }
             case None =>
                 Future.successful(Unauthorized(Json.toJson(StatusResponse(StatusCode.UNAUTHORIZED.code, StatusCode.UNAUTHORIZED.message))))
         }
