+++ restaurant-remote-reception/conf/application.conf
@@ -13,14 +13,18 @@ play.filters.hosts {
 
 slick.dbs.default.driver="slick.driver.MySQLDriver$"
 slick.dbs.default.db.driver = "com.mysql.jdbc.Driver"
-slick.dbs.default.db.url=${?MYSQL_DATABASE_URL}
-slick.dbs.default.db.user=${?MYSQL_DATABASE_USER}
-slick.dbs.default.db.password=${?MYSQL_DATABASE_PWD}
+//slick.dbs.default.db.url=${?MYSQL_DATABASE_URL}
+slick.dbs.default.db.url="jdbc:mysql://localhost:3306/restaurant_remote_reception"
+slick.dbs.default.db.user="root"
+//slick.dbs.default.db.user=${?MYSQL_DATABASE_USER}
+//slick.dbs.default.db.password=${?MYSQL_DATABASE_PWD}
+slick.dbs.default.db.password="mysql"
 
 play.evolutions.enabled=true
 
 play.filters.cors {
-  allowedOrigins = [${?ALLOWED_ORIGIN}]
+//  allowedOrigins = [${?ALLOWED_ORIGIN}]
+  allowedOrigins = ["localhost:4200"]
   allowedHttpMethods = ["GET", "POST", "DELETE", "OPTIONS", "PUT"]
   allowedHttpHeaders = ["Content-Type", "Accept"]
 }
